moved helper functions from line_tools.py to utils.py
