+++ spectools/utils.py
@@ -1,0 +1,9 @@
+import numpy as np
+
+def find_nearest(array, value):
+    array = np.asarray(array)
+    idx = (np.abs(array - value)).argmin()
+    return array[idx],idx
+
+def nlayers(n):
+    return 2*n-1
